refactor: restructure publishers module, fix routing and update stats API fields
