+++ src/app/app.routes.ts
@@ -71,37 +71,27 @@ export const routes: Routes = [
   },
   {
     path: 'quranic-cms',
+    loadComponent: () =>
+      import('./features/quranic-cms/quranic-cms.page').then((m) => m.QuranicCmsPage),
+    canActivate: [authGuard],
+    data: { hideHeader: true, fullWidth: true },
     children: [
+      {
+        path: '',
+        loadChildren: () =>
+          import('./features/quranic-cms/quranic-cms.routes').then((m) => m.quranicCmsRoutes),
+      },
       {
         path: 'publishers',
         loadChildren: () =>
-          import('./features/quranic-cms/publishers/publishers/publishers.routes').then(
+          import('./features/quranic-cms/publishers/publishers.routes').then(
             (m) => m.publishersRoutes
           ),
-      },
-      {
-        path: '',
-        redirectTo: 'publishers',
-        pathMatch: 'full',
+        canActivate: [publisherHostGuard],
       },
     ],
   },
-  {
-    path: 'publishers',
-    loadComponent: () =>
-      import('./features/publishers/pages/publishers/publishers.page').then(
-        (m) => m.PublishersPage
-      ),
-    canActivate: [publisherHostGuard], // Restrict access for publisher hosts
-  },
-  {
-    path: 'publisher/:id',
-    loadComponent: () =>
-      import('./features/publishers/pages/publisher-details/publisher-details.page').then(
-        (m) => m.PublisherDetailsPage
-      ),
-    canActivate: [publisherHostGuard], // Restrict access for publisher hosts
-  },
+
   {
     path: 'license/:id',
     loadComponent: () =>
