--- src/app/features/quranic-cms/publishers/publishers-layout.component.html
@@ -1,6 +1,0 @@
-<app-qcms-section-layout
-  title="الناشرون والمصادر"
-  description="إدارة الناشرين والمؤلفين والمصادر المستخدمة في المشاريع القرآنية"
-  [basePath]="basePath"
-  [tabs]="subTabs"
-/>
