--- src/app/features/quranic-cms/publishers/publishers-layout.component.less
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
