+++ src/app/features/quranic-cms/publishers/publishers-layout.component.ts
@@ -1,25 +1,86 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { SectionLayoutComponent } from '../components/section-layout/section-layout.component';
+import { RouterModule } from '@angular/router';
 
 @Component({
   selector: 'app-publishers-layout',
   standalone: true,
-  imports: [SectionLayoutComponent],
+  imports: [CommonModule, RouterModule],
   template: `
-    <app-section-layout
-      title="الناشرون والمصادر"
-      description="إدارة الناشرين والمؤلفين والمصادر المستخدمة في المشاريع القرآنية"
-      [basePath]="basePath"
-      [tabs]="tabs"
-    />
-  `,
-})
-export class PublishersLayoutComponent {
-  readonly basePath = '/quranic-cms/publishers';
+    <div class="publishers-layout">
+      <nav class="custom-nav-container">
+        <div class="nav-pill-wrapper">
+          <a routerLink="/quranic-cms/publishers/sources" routerLinkActive="active" class="nav-item"
+            >المصادر</a
+          >
 
-  readonly tabs = [
-    { id: 'publishers', path: 'publishers', label: 'الناشرون' },
-    { id: 'authors', path: 'authors', label: 'المؤلفون' },
-    { id: 'sources', path: 'sources', label: 'المصادر' },
-  ];
-}
+          <a routerLink="/quranic-cms/publishers/authors" routerLinkActive="active" class="nav-item"
+            >المؤلفون</a
+          >
+
+          <a
+            routerLink="/quranic-cms/publishers"
+            [routerLinkActiveOptions]="{ exact: true }"
+            routerLinkActive="active"
+            class="nav-item"
+            >الناشرون</a
+          >
+        </div>
+      </nav>
+
+      <main class="content-area">
+        <router-outlet></router-outlet>
+      </main>
+    </div>
+  `,
+  styles: [
+    `
+      .publishers-layout {
+        background: white;
+        margin-top: 45px;
+        border-radius: 15px;
+        border: solid #f0f2f5 3px;
+        padding: 20px 0;
+      }
+
+      .custom-nav-container {
+        display: flex;
+        justify-content: center;
+        margin-bottom: 30px;
+      }
+
+      .nav-pill-wrapper {
+        background-color: #f0f2f5;
+        border-radius: 50px;
+        padding: 4px;
+        display: inline-flex;
+        border: 1px solid #e8e8e8;
+        min-width: 400px;
+        justify-content: space-between;
+      }
+
+      .nav-item {
+        padding: 8px 35px;
+        border-radius: 50px;
+        text-decoration: none;
+        color: #8c8c8c;
+        font-size: 14px;
+        transition: all 0.3s ease;
+        text-align: center;
+        flex: 1;
+      }
+
+      .active {
+        background-color: #ffffff;
+        color: #262626 !important;
+        box-shadow: 0 2px 8px rgba(0, 0, 0, 0.06);
+        font-weight: 500;
+      }
+
+      .content-area {
+        padding: 0 24px;
+      }
+    `,
+  ],
+})
+export class PublishersLayoutComponent {}
