+++ src/app/features/quranic-cms/publishers/publishers.routes.ts
@@ -1,35 +1,31 @@
 import { Routes } from '@angular/router';
 import { PublishersLayoutComponent } from './publishers-layout.component';
+import { PublishersComponent } from './publishers.component';
 
 export const publishersRoutes: Routes = [
   {
     path: '',
     component: PublishersLayoutComponent,
     children: [
-      { path: '', redirectTo: 'publishers', pathMatch: 'full' },
       {
-        path: 'publishers',
-        loadComponent: () =>
-          import('../components/coming-soon/coming-soon.component').then(
-            (m) => m.ComingSoonComponent
-          ),
-        data: { emoji: '📰' },
+        path: '',
+        component: PublishersComponent,
       },
       {
         path: 'authors',
         loadComponent: () =>
-          import('../components/coming-soon/coming-soon.component').then(
-            (m) => m.ComingSoonComponent
+          import('../../../shared/components/empty-placeholder/empty-placeholder.component').then(
+            (m) => m.EmptyPlaceholderComponent
           ),
-        data: { emoji: '✍️' },
+        data: { title: 'المؤلفون' },
       },
       {
         path: 'sources',
         loadComponent: () =>
-          import('../components/coming-soon/coming-soon.component').then(
-            (m) => m.ComingSoonComponent
+          import('../../../shared/components/empty-placeholder/empty-placeholder.component').then(
+            (m) => m.EmptyPlaceholderComponent
           ),
-        data: { emoji: '🗂️' },
+        data: { title: 'المصادر' },
       },
     ],
   },
