+++ src/app/features/quranic-cms/publishers/services/publishers.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { environment } from '../../../../../../environments/environment';
+import { environment } from '../../../../../environments/environment';
 import { Publisher } from '../models/publishers-stats.models';
 
 @Injectable({
